Remove isAble method in OrderController in favor of Gates/Policies

+++ app/Http/Controllers/Api/V1/OrderController.php
@@ -10,11 +10,10 @@ use App\Http\Resources\V1\OrderResource;
 use App\Models\Order;
 use App\Policies\V1\OrderPolicy;
 use App\Services\V1\OrdersService;
-use Illuminate\Auth\Access\AuthorizationException;
-use Illuminate\Database\Eloquent\ModelNotFoundException;
 use Illuminate\Database\QueryException;
 use Illuminate\Http\Response;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 use Throwable;
 
 class OrderController extends ApiController
@@ -30,6 +29,8 @@ class OrderController extends ApiController
      */
     public function index(OrderFilter $filter)
     {
+        Gate::authorize('view-any', Order::class);
+
         return new OrderCollection(
             Order::filter($filter)->paginate()
         );
@@ -40,7 +41,7 @@ class OrderController extends ApiController
      */
     public function show(Order $order)
     {
-        $this->isAble('view', $order); // policy
+        Gate::authorize('view', $order);
 
         if ($this->include('user')) {
             $order->load('user');
@@ -56,6 +57,8 @@ class OrderController extends ApiController
      */
     public function store(StoreOrderRequest $request)
     {
+        Gate::authorize('create', Order::class);
+
         try {
             $order = $this->orderService->createOrderHandleProducts($request);
 
@@ -77,8 +80,9 @@ class OrderController extends ApiController
      */
     public function update(UpdateOrderRequest $request, Order $order)
     {
+        Gate::authorize('update', $order);
+
         try {
-            $this->isAble('update', $order); // policy
             $this->orderService->updateOrderHandleProducts($request, $order);
 
             return response()->json(new OrderResource($order), Response::HTTP_OK);
@@ -98,7 +102,7 @@ class OrderController extends ApiController
      */
     public function destroy(Order $order)
     {
-        $this->isAble('delete', $order); // policy
+        Gate::authorize('delete', $order);
         $this->orderService->deleteOrderHandleProducts($order);
 
         return $this->responseSuccess('Order deleted successfully');
